refactor: move currentColor to renderableView
